% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{fullTab.probs}
\alias{fullTab.probs}
\title{Calculates full table probabilities}
\usage{
fullTab.probs(mix.probs, class.probs, vals, sample = NULL)
}
\arguments{
\item{mix.probs}{Matrix of mixing probability}

\item{class.probs}{Array of class probability for latent classes}

\item{vals}{The full table cell that should be calculated}

\item{sample}{What sample should the full cell probability be calculated for? `NULL` defaults to 1.}
}
\value{
A full table probability for a given cell
}
\description{
Calculates full table probabilities
}
\keyword{internal}
