% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getProbs}
\alias{getProbs}
\title{Helper function to calculate two-way probabilities}
\usage{
getProbs(comb, cols, samp, G.used, class.probs)
}
\description{
Helper function to calculate two-way probabilities
}
\keyword{internal}
