% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFunctions.R
\name{lap.lik}
\alias{lap.lik}
\title{Calculate the likelihood for the discrete Laplace distribution}
\usage{
lap.lik(data, prop, E = -1/6)
}
\arguments{
\item{data}{A vector of counts}

\item{prop}{A vector representing the mean of the counts}

\item{E}{The transformed epsilon to match the geometric distribution}
}
\value{
The value of the likelihood evaulated at `data`
}
\description{
Calculate the likelihood for the discrete Laplace distribution
}
