% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFunctions.R
\name{likelihood.vec}
\alias{likelihood.vec}
\title{The binomial likelihood function for a specific count and probability vector}
\usage{
likelihood.vec(
  counts,
  samp.size,
  G,
  sample,
  counts.mat,
  mix.probs,
  comp.probs,
  index = NULL,
  par = FALSE,
  clust
)
}
\arguments{
\item{counts}{A vector of noisy counts}

\item{samp.size}{Sample size}

\item{G}{The number of latent classes}

\item{sample}{What sample in the chain should the likelihood be evaulated for?}

\item{counts.mat}{A matrix encoding what counts correspond to what variables}

\item{mix.probs}{A # of samples * G matrix of mixing probabilities}

\item{comp.probs}{An array of the component probabilities for each latent class and run}

\item{index}{Defaults to NULL. If a specific value, will only calculate the likelihood corresponding to the count in that index.}

\item{par}{Boolean. Do we parallelize the code?}

\item{clust}{If paralleized, what cluster do we use}
}
\value{
The evaluated likelihood
}
\description{
The binomial likelihood function for a specific count and probability vector
}
