% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFunctions.R
\name{mcmc.sampler}
\alias{mcmc.sampler}
\title{The main sampling function for the Bayesian Latent Class approach}
\usage{
mcmc.sampler(
  counts,
  counts.mat,
  eps = 1,
  P,
  G = 5,
  nsamples = 2000,
  samp.size,
  E.var = NULL,
  .calculateFullTabProbs = FALSE,
  .PiTuningParam = 0.5,
  .PsiTuningParam = 10,
  .methodPi = "normal",
  cl = NULL
)
}
\arguments{
\item{counts}{A vector of the true counts}

\item{counts.mat}{A matrix encoding what counts correspond to what variables}

\item{eps}{The value for epsilon}

\item{P}{The number of variables}

\item{G}{The number of latent classes}

\item{nsamples}{The number of samples to be drawn}

\item{samp.size}{Sample size}

\item{E.var}{Defaults to NULL. A value can be supplied to change the tuning parameter when sampling for \eqn{\Tilde{M}}}

\item{.calculateFullTabProbs}{A Boolean. Do we calculate the full table probabilities? It can be expensive to do so when dimension or the number of latent classes is large.}

\item{.PiTuningParam}{A small constant used to tune the proposal distribution of \eqn{\pi}.}

\item{.PsiTuningParam}{A small constant used to tune the proposal distribution of \eqn{\psi}.}

\item{.methodPi}{What proposal distribution ("normal" or "dirichlet") should be used for sampling for pi?}

\item{cl}{Cluster used if parallelized}
}
\value{
A list containing samples of \eqn{\Tilde{M}}, \eqn{\pi}, and \eqn{\psi} as well as marginal probabilities and full table probabilities (if desired). Also contains acceptance rates for all of the parameters.
}
\description{
The main sampling function for the Bayesian Latent Class approach
}
