% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{row.prob.vec}
\alias{row.prob.vec}
\title{Helper function to calculate two-way probabilities}
\usage{
row.prob.vec(combs.full, mix.probs, class.probs, P, sample, G)
}
\arguments{
\item{combs.full}{A matrix of all combinations of the other variables that we need to sum over. Note: for two-way probabilities, two variables remain fixed, must iterate over all combinations of the others.}

\item{mix.probs}{Matrix of mixing probability}

\item{class.probs}{Array of class probability for latent classes}

\item{P}{The dimension of the table}

\item{sample}{What sample should the marginal probability be calculated for?}

\item{G}{The number of latent classes}
}
\value{
A component of the marginal probability
}
\description{
Helper function to calculate two-way probabilities
}
\keyword{internal}
