% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{twoWay.probs.vectorized}
\alias{twoWay.probs.vectorized}
\title{Vectorized function to count two-way marginal probabilities for a full vector or single count}
\usage{
twoWay.probs.vectorized(
  counts.vec,
  mix.probs,
  class.probs,
  index1 = NULL,
  index2 = NULL,
  val1 = NULL,
  val2 = NULL,
  sample = 1,
  G
)
}
\arguments{
\item{counts.vec}{A vector signifying chick combination is desired. Should only have two enteries and the rest should be `NA`}

\item{mix.probs}{Matrix of mixing probability}

\item{class.probs}{Array of class probability for latent classes}

\item{index1}{Index of first marginal cell}

\item{index2}{Index of second marginal cell. Must be different from index1.}

\item{val1}{Value corresponding to the first marginal cell.}

\item{val2}{Value corresponding to the second marginal cell.}

\item{sample}{What sample should the marginal probability be calculated for?}

\item{G}{The number of latent classes}
}
\value{
A marginal probability for a cell
}
\description{
Vectorized function to count two-way marginal probabilities for a full vector or single count
}
\keyword{internal}
